+++ tests/test_source_field.py
@@ -228,6 +228,13 @@ class StraightFieldTests(test.TestCase):
         self.assertEqual(len(obj), 1)
         self.assertEqual(obj[0].chars_upper, "AAA")
 
+    async def test_values_by_fk(self):
+        obj1 = await self.model.create(chars="aaa")
+        await self.model.create(chars="bbb", fk=obj1)
+
+        obj = await self.model.filter(chars="bbb").values("fk__chars")
+        self.assertEqual(obj, [{"fk__chars": "aaa"}])
+
 
 class SourceFieldTests(StraightFieldTests):
     def setUp(self) -> None: