+++ tortoise/filters.py
@@ -3,8 +3,7 @@ from functools import partial
 from typing import TYPE_CHECKING, Any, Dict, Iterable, Optional, Tuple
 
 from pypika import Table
-from pypika.enums import SqlTypes
-from pypika.functions import Cast, Upper
+from pypika.functions import Upper
 from pypika.terms import BasicCriterion, Criterion, Equality, Term, ValueWrapper
 
 from tortoise.fields import Field
@@ -48,12 +47,14 @@ def string_encoder(value: Any, instance: "Model", field: Field) -> str:
 def is_in(field: Term, value: Any) -> Criterion:
     if value:
         return field.isin(value)
+    # SQL has no False, so we return 1=0
     return BasicCriterion(Equality.eq, ValueWrapper(1), ValueWrapper(0))
 
 
 def not_in(field: Term, value: Any) -> Criterion:
     if value:
         return field.notin(value) | field.isnull()
+    # SQL has no True, so we return 1=1
     return BasicCriterion(Equality.eq, ValueWrapper(1), ValueWrapper(1))
 
 
@@ -78,31 +79,31 @@ def not_null(field: Term, value: Any) -> Criterion:
 
 
 def contains(field: Term, value: str) -> Criterion:
-    return Cast(field, SqlTypes.VARCHAR).like(f"%{value}%")
+    return field.like(f"%{value}%")
 
 
 def starts_with(field: Term, value: str) -> Criterion:
-    return Cast(field, SqlTypes.VARCHAR).like(f"{value}%")
+    return field.like(f"{value}%")
 
 
 def ends_with(field: Term, value: str) -> Criterion:
-    return Cast(field, SqlTypes.VARCHAR).like(f"%{value}")
+    return field.like(f"%{value}")
 
 
 def insensitive_exact(field: Term, value: str) -> Criterion:
-    return Upper(Cast(field, SqlTypes.VARCHAR)).eq(Upper(f"{value}"))
+    return Upper(field).eq(Upper(f"{value}"))
 
 
 def insensitive_contains(field: Term, value: str) -> Criterion:
-    return Upper(Cast(field, SqlTypes.VARCHAR)).like(Upper(f"%{value}%"))
+    return Upper(field).like(Upper(f"%{value}%"))
 
 
 def insensitive_starts_with(field: Term, value: str) -> Criterion:
-    return Upper(Cast(field, SqlTypes.VARCHAR)).like(Upper(f"{value}%"))
+    return Upper(field).like(Upper(f"{value}%"))
 
 
 def insensitive_ends_with(field: Term, value: str) -> Criterion:
-    return Upper(Cast(field, SqlTypes.VARCHAR)).like(Upper(f"%{value}"))
+    return Upper(field).like(Upper(f"%{value}"))
 
 
 ##############################################################################