+++ CHANGELOG.rst
@@ -30,6 +30,9 @@ Changelog
   * If you do a ``<model>.save(update_fields=[...])`` and one of the fields in ``update_fields`` was not in the ``.only(...)``,
     then ``IncompleteInstanceError`` as that field is not available to be updated.
 
+- Fixed bad SQL generation when doing a ``.values()`` query over a Foreign Key
+- Added `<model>.update_from_dict({...})` that will mass update values safely from a dictionary
+
 0.16.5
 ------
 * Moved ``Tortoise.describe_model(<MODEL>, ...)`` to ``<MODEL>.describe(...)``